Student Side API's finished
